requirements.txt
Django==5.0.2
python-dotenv==1.0.1
groq==0.4.2
whitenoise==6.6.0
requests>=2.31.0
